requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
